requirements.txt
flask
mysql-connector-python
python-dotenv
gunicorn
requests
Flask-Session
pytz
email-validator
phonenumbers